% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylSearch.R
\name{getMethylOverlap}
\alias{getMethylOverlap}
\title{Get methylation data overlapping with ChIP-SEQ peaks}
\usage{
getMethylOverlap(chipPath, methylPath)
}
\arguments{
\item{chipPath}{the path to a BED6 or greater file containing processed
ChIP-seq peaks. If data is not stranded, a period should be specified
in the strand column.}

\item{methylPath}{the path to a BED9 file with methylation data similar
to those available on the UCSC website, where the ninth column
specifies the percentage of reads that were methylated:
http://hgdownload.cse.ucsc.edu/goldenPath/hg19/encodeDCC/wgEncodeHaibMethylRrbs/}
}
\value{
a dataframe where the first column specifies
a chromosome, the second column specifies an end position
on the genome, the sixth column represents a strand or . if
this is not specified, and the last (ninth) column specifies
the percentage of reads that were methylated.
}
\description{
Given paths to a BED file with ChIP-SEQ peaks and a BED file
with methylation data, return a dataframe containing methylation
data for sites within ChIP-SEQ peaks.
}
\examples{
system.file("extdata", "MAZ_very_small_test.bed", package = "ChIPAnalyzer")
system.file("extdata", "HcfUMethylData.bed", package = "ChIPAnalyzer")
overlap <- getMethylOverlap("MAZ_very_small_test.bed", "HcfUMethylData.bed")

}
\references{
H. Pagès, M. Lawrence and P. Aboyoun (2020). S4Vectors:
Foundation of vector-like and list-like containers in
Bioconductor. R package version 0.24.4.

Lawrence M, Huber W, Pag\`es H, Aboyoun P, Carlson M, et al.
(2013) Software for Computing and Annotating Genomic Ranges.
PLoS Comput Biol 9(8): e1003118.
doi:10.1371/journal.pcbi.1003118
}
