% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadCalc.R
\name{getQuadCoveragePercentage}
\alias{getQuadCoveragePercentage}
\title{Get the percentage abundance of quadruplexes at positons across multiple
sequences of equal length}
\usage{
getQuadCoveragePercentage(quadMatrix)
}
\arguments{
\item{quadMatrix}{a matrix in that format returned from getQuadMatrix,
with each row representing a different sequence a columns representing
positions, with a 1 indicated presence of a quadruplex and 0 otherwise

reports <- findQuads(bedPath = "BRD4_reduced.bed", seqWidth = 75, assemblyVersion = "hg19")
qMatrix <- getQuadMatrix(quadReports = reports, seqWidth = 75)
quadCoveragePercentage <- getQuadCoveragePercentage(quadMatrix = qMatrix)@return a vector of percentages indicating the percentage of rows in
quadmatrix that there was a 1(quadruplex) at the position represented
by that column}
}
\description{
Using the matrix returned by getQuadMatrix, calculate the percentage abundance
quadruplexes at positions dictated by the columns of quadMatrix
}
