% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plotQuadPosition}
\alias{plotQuadPosition}
\title{Plot the abundance of Predicted G-quadruplexes
Relative to a Central Postiion}
\usage{
plotQuadPosition(featurePercentages, title)
}
\arguments{
\item{featurePercentages}{a numeric vector containing percentages such as
a vector returned by getQuadCoveragePercentage()}

\item{title}{a string specifying the title of the plot}
}
\value{
returns a plot of G-Quadruplex Abundance
}
\description{
Using a vector returned by getQuadCoveragePercentage(),
plot the percentage of peaks that have G-quadruplexes across
positions in sequences around the ChIP-SEQ peaks
}
\examples{
path <- system.file("extdata", "MAZ_very_small_test.bed", package = "ChIPAnalyzer")
reports <- findQuads(bedPath = path,
 seqWidth = 200,
  assemblyVersion = "hg19")
qMatrix <- getQuadMatrix(quadReports = reports)
quadCoveragePercentage <- getQuadCoveragePercentage(quadMatrix = qMatrix)
plotQuadPosition(quadCoveragePercentage, "MAZ G-Quad Abundance")

}
