% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runChIPAnalyzer.R
\name{runChIPAnalyzer}
\alias{runChIPAnalyzer}
\title{Run the Shiny App for ChIPAnalyzer}
\usage{
runChIPAnalyzer()
}
\description{
This function launches the Shiny GUI for ChIPAnalyzer
}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
