% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadCalc.R
\name{findQuads}
\alias{findQuads}
\title{Generate Reports on Potential G-quadruplex forming Strings
near ChIP peaks using pqsfinder}
\usage{
findQuads(bedPath, seqWidth, assemblyVersion = "hg19")
}
\arguments{
\item{bedPath}{the path to a BED6 or greater file containing processed
ChIP-seq peaks. If data is not stranded, a period should be specified
in the strand column.}

\item{seqWidth}{the width of the sequence around the centre of each peak that
should be searched for quadruplexes.}

\item{assemblyVersion}{the genome assembly to retrive sequences from.
Must be one of "hg19", "hg38", "mm9" or "mm10"}
}
\value{
returns a list of reports on potential quadruplex forming strings
generated by pqsfinder::pqsfinder().
}
\description{
Searches for G-quadruplexes in DNA sequences around the ChIP peaks from the BED6 or
greater file specified throught the function argument bedPath, and returns a list
of reports for each sequence generated by the pqsfinder package.
}
\examples{
findQuads(bedPath = "MAZ_high_score.bed", seqWidth = 200, assemblyVersion = "hg19")
}
