% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadCalc.R
\name{getQuadCoveragePercentage}
\alias{getQuadCoveragePercentage}
\title{Get the percentage abundance of quadruplexes at positons across multiple
sequences of equal length}
\usage{
getQuadCoveragePercentage(quadMatrix)
}
\arguments{
\item{quadMatrix}{a matrix in that format returned from getQuadMatrix,
with each row representing a different sequence a columns representing
positions, with a 1 indicated presence of a quadruplex and 0 otherwise}
}
\value{
a vector of percentages indicating the percentage of rows in
quadmatrix that there was a 1(quadruplex) at the position represented
by that column
}
\description{
Using the matrix returned by getQuadMatrix, calculate the percentage abundance
quadruplexes at positions dictated by the columns of quadMatrix
}
\examples{
reports <- findQuads(bedPath = "MAZ_high_score.bed",
 seqWidth = 200,
 assemblyVersion = "hg19")
qMatrix <- getQuadMatrix(quadReports = reports, maxOnly = TRUE)
quadCoveragePercentage <-
 getQuadCoveragePercentage(quadMatrix = qMatrix)

}
