% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plotMethylPercentage}
\alias{plotMethylPercentage}
\title{Generate a Pie Chart of The percentage of ChIP peak nucleotides that
are methylated}
\usage{
plotMethylPercentage(methylOverlapData)
}
\arguments{
\item{methylOverlapData}{a data frame with a coverage column specifying
the percentage of reads methylated. This column must be named coverage}
}
\description{
Using the dataframe of methylation data for ChIP peak nucleotides
generated by getMethylOverlap, create a pie chart showing the
percentage of peak sites that are methylated
}
